chore: read and write from client cache
