+++ src/graphql/client.ts
@@ -7,8 +7,7 @@ import { InMemoryCache } from 'apollo-cache-inmemory'
 import { HttpLink } from 'apollo-link-http'
 import { getMainDefinition } from 'apollo-utilities'
 import { getCookie, removeCookie, redirect } from '../utils/helpers'
-import gql from 'graphql-tag'
-
+import resolvers from './resolvers'
 const httpLink = new HttpLink({
   uri: process.env.REACT_APP_GRAPHQL_URI,
 })
@@ -87,45 +86,14 @@ const terminatingLink = split(
   authLink.concat(httpLink)
 )
 
-const ADD_EXPERIENCE = gql`
-  mutation addExperience($experience: ExperienceInput!) {
-    addExperience(experience: $experience) @client {
-      name
-    }
-  }
-`
-
-const GET_EXPERIENCES = gql`
-  {
-    experiences @client {
-      experience
-    }
-  }
-`
 const apolloClient = new ApolloClient({
   link: ApolloLink.from([onError(handleErrors), terminatingLink]),
   cache,
-  resolvers: {
-    Mutation: {
-      addExperience: (
-        experience: any,
-        variables: any,
-        { cache }: { cache: any }
-      ): any => {
-        const { experiences } = cache.readQuery({ query: GET_EXPERIENCES })
-        console.log('experiences:', experiences)
-        experiences.push(variables)
-        console.log('adding', variables)
-
-        return cache.writeQuery({ query: GET_EXPERIENCES, data: experiences })
-      },
-    },
-  },
+  resolvers,
 })
 const initialState = {
   name: 'test',
-  experiences: [
-  ],
+  experiences: [],
 }
 
 cache.writeData({
